+++ components/PolarAreaChart.js
@@ -41,6 +41,7 @@ export default function PolarAreaChart({
     return {
       key,
       value,
+      percent,
       d: `
         M ${size/2} ${size/2}
         L ${start.x} ${start.y}
@@ -61,15 +62,41 @@ export default function PolarAreaChart({
               r={radii / arr.length * (i+1)}
               cx={size/2}
               cy={size/2}
+              fill='none'
+              stroke='var(--gps-border-color)'
             />
           ))}
-          {items.map((item, i) => (
-            <path
-              key={item.key}
-              d={item.d}
-              fill={item.color}
-            />
-          ))}
+          {items.length === 1 ? (
+            items.map((item, i) => (
+              <circle
+                key={item.key}
+                fill={item.color}
+                r={size/2}
+                cx={size/2}
+                cy={size/2}
+                data-tip={`
+                  ${item.key}: ${item.value}
+                  <br/>
+                  <strong>${Math.round(item.percent * 100)}%</strong>
+                `}
+                data-html={true}
+              />
+            ))
+          ) : (
+            items.map((item, i) => (
+              <path
+                key={item.key}
+                d={item.d}
+                fill={item.color}
+                data-tip={`
+                  ${item.key}: ${item.value}
+                  <br/>
+                  <strong>${Math.round(item.percent * 100)}%</strong>
+                `}
+                data-html={true}
+              />
+            ))
+          )}
         </svg>
       </div>
       <div className='info'>
@@ -92,10 +119,6 @@ export default function PolarAreaChart({
         }
         svg {
           transform: rotate(90deg);
-        }
-        circle {
-          fill: none;
-          stroke: var(--gps-border-color);
         }
         path {
           stroke: var(--color-background);