+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+(feature/tooltips)
+#### Added
+- user: organizations: tooltip to show org name
+- user: CalendarLine (chart): tooltip to show grouped contributions per week
+- components: PolarAreaChart: tooltip to show percentage
+- components: PieChart: tooltip to show per percentage
+
 (feature/area-chart-update)
 #### Fixed
 - components: PolarAreaChart: handle case for 1 entry