+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.7.0] - 2021-04-17
+#### Added
+- dependencies: add react-tooltip
+- user
+  - organizations: tooltip to show org name
+  - CalendarLine (chart): tooltip to show grouped contributions per week
+- components
+  - PolarAreaChart: tooltip to show percentage
+  - PieChart: tooltip to show per percentage
+#### Fixed
+- components: PolarAreaChart: handle case for 1 entry
+
 ## [0.6.0] - 2021-04-15
 #### Added
 - components: polar area chart
@@ -67,6 +79,7 @@ Initial release
   - languages per repo
   - commits per repo (top 10)
 
+[0.7.0]: https://github.com/noeldelgado/gh-profile-stats/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/noeldelgado/gh-profile-stats/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/noeldelgado/gh-profile-stats/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/noeldelgado/gh-profile-stats/compare/v0.3.0...v0.4.0